requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.0.0
psycopg2>=2.8

# Test requirements
pytest-django==2.9.1
pytest==2.8.5
pytest-cov==2.2.0

# wheel for PyPI installs
wheel==0.24.0